Do not fetch a new word on final page
State loads should not have non-deterministic behaviour. In restart()
which is triggered in the loadState() function a new word was randomly
picked, meaning that each time you go back or forward to the Grattis
page you would pick a new word as the next which messed up the
succeeding state. Now new word is picked until you click restart and
that is pushed into the state of the next page together with the full
list of words again (word_data is refilled).